@@ -1,5 +1,5 @@
 {
-  "name": "golemgrid-frontend",
+  "name": "golemgrid",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,21 +8,29 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "bootstrap": "^5.0.1",
-    "bootstrap-vue": "^2.21.2",
+    "@headlessui/vue": "^1.3.0",
+    "@heroicons/vue": "^1.0.1",
+    "@tailwindcss/forms": "^0.3.3",
+    "@tailwindcss/postcss7-compat": "^2.0.2",
+    "autoprefixer": "^9",
+    "axios": "^0.21.1",
     "core-js": "^3.6.5",
-    "vue": "^2.6.14",
-    "vue-router": "^3.2.0"
+    "postcss": "^7",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.2",
+    "vue": "^3.0.0",
+    "vue-number-animation": "^1.0.5",
+    "vue-router": "^4.0.0-0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0",
+    "vue-cli-plugin-tailwind": "~2.0.6"
   },
   "eslintConfig": {
     "root": true,
@@ -30,7 +38,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
