@@ -1,29 +1,22 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Blender from '../views/Blender.vue'
-
-Vue.use(VueRouter)
+import { createRouter, createWebHistory } from "vue-router"
+import Home from "../views/Home.vue"
+import TaskInfo from "../views/TaskInfo.vue"
 
 const routes = [
   {
-    path: '/',
-    name: 'Blender',
-    component: Blender,
+    path: "/",
+    name: "Home",
+    component: Home,
   },
   {
-    path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () =>
-      import(/* webpackChunkName: "about" */ '../views/About.vue'),
+    path: "/task/:id",
+    name: "taskinfo",
+    component: TaskInfo,
   },
 ]
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes,
 })
 
